Organized some of the general chatterbot test cases.
