+++ tests/base_case.py
@@ -3,7 +3,7 @@ from chatterbot import ChatBot
 import os
 
 
-class UntrainedChatBotTestCase(TestCase):
+class ChatBotTestCase(TestCase):
 
     def setUp(self):
         self.test_data_directory = 'test_data'
@@ -39,35 +39,4 @@ class UntrainedChatBotTestCase(TestCase):
         """
         self.chatbot.storage.drop()
         self.remove_test_data()
-
-
-class ChatBotTestCase(UntrainedChatBotTestCase):
-
-    def setUp(self):
-        super(ChatBotTestCase, self).setUp()
-        """
-        Set up a database for testing.
-        """
-        data1 = [
-            "african or european?",
-            "Huh? I... I don't know that.",
-            "How do you know so much about swallows?"
-        ]
-
-        data2 = [
-            "Siri is adorable",
-            "Who is Seri?",
-            "Siri is my cat"
-        ]
-
-        data3 = [
-            "What... is your quest?",
-            "To seek the Holy Grail.",
-            "What... is your favourite colour?",
-            "Blue."
-        ]
-
-        self.chatbot.train(data1)
-        self.chatbot.train(data2)
-        self.chatbot.train(data3)
